Basic UI mockup of recipe notebook
still need to add implementation for selecting and changing recipes.

this UI mockup can be used for both menus (the one for the start of the day and the one the player can see at any point)

need to add boolean to hide bake button for second menu (the one the player can view at any time).

maybe also add UI element to show which recipe has been selected since 2 recipes need to be selected and they can be on different pages