Interactables is done! (I think)
Added a scene (player_interact_test), where the player can be moved around to interact with two nodes. The player gets the interactable node that is closest to the player, and when the player presses the interact button (E in the GDD, space in the main branch), the "_interact" function of the node is called!