Better Temporary Phone UI
Interface of the phone with temporary art for the apps, they've all been turned into texture buttons so the app art will be rounded squares. The phone has a simple temporary opening animation and a script however nothing was added to the script. Plan for the script is to catch the signal of the invisible texture button being pressed to then start the open_phone animation.