SceneTransition: Change animation to anchors & make color changeable
Now the animation is based on anchors, which should (?) be more precise.

More usefully, though, the color is now split into modulate and
color, so the actual color of the transition can be changed by
changing the ColorRect's color (if we don't want as strong of pink,
as Justin was saying).