Initial / experimental player palette swap
Note the important notes:
- Player textures must not be drawn with antialiasing
- They must be drawn with exactly the same palette
- Any Godot nodes using the shader must have texture filter set to Nearest

Besides that, it is as simple as simply adding the player_palette
shader / material to the nodes that need the palette swap. The
colors themselves are global uniforms so they are the same across
the entire project.

One other note:
I think the max texture size in Godot in 16384, although I should
look this up. If that's true, then we need to probably crop each
frame inside the animations so they fit.

(It would really be ideal if we "just" did any relevant skeletal
animation directly inside Godot -- the shader would continue to
work, but the animation would likely be somewhat smoother (?) and
at the very least would be much more GPU efficient without needing
to bake the same textures many times into animation frames).