Add enum for foodType in food_data
This change feels majorly scuffed, but it was the best way that I could conceive detecting the type of food being done.
I was originally trying to hack it using the recipe display name, but it would have immediately broken if someone changed a name.
The enum is technically a duplicate of the one from FoodItem, but FoodItem it completely unused outside of lunch rush.