Added and ALMOST completed salty date 2 dialogue
I created a timeline for salty date 2 and ported the script into dialogic. Please note that the script did not include exposition at the beginning, so this still needs to be added. There is also a part of the script where salty refers to the player by name, but as of right now he just says "[Player]". The player name variable still needs to be passed in.